requirements.txt
chromadb == 0.5.3
Flask==3.0.3
Flask-Cors==4.0.1
langchain==0.2.2
langchain-community==0.2.2
langchain-core==0.2.4
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
openai==1.31.0
pypdf==4.3.0
python-dotenv==1.0.1
faiss-cpu==1.8.0
pdfplumber



